refactor: API_METHOD_TYPE, API_STATUS의 패키지 이동
