+++ fourtoon-cookie/src/apis/diary.ts
@@ -3,7 +3,7 @@ import type { DiaryCreatedResponse, DiaryPatchFavoriteRequest, DiarySaveRequest,
 import { requestApi } from "./api";
 import { LocalDate } from "@js-joda/core";
 import { ApiError } from "../types/error/ApiError";
-import { API_METHOD_TYPE, API_STATUS } from "../constants/api";
+import { API_METHOD_TYPE, API_STATUS } from "../types/api";
 
 export const getDiary = async (diaryId: number): Promise<Diary> => {
     const response = await requestApi(`/diary/${diaryId}`, API_METHOD_TYPE.GET);
